requirements.txt
pillow>=9.0.0
mss>=6.0.0
pygetwindow>=0.0.9
pytesseract>=0.3.8
pyserial>=3.5
opencv-python>=4.5.0
